Adapt mhlo.conv to Linalg on buffers patterns to be on tensors. (#4748)
Lowers a conv op to a static rank conv op. The general conv op only
supports on memrefs. This limits the support for conv op to specific
data layout and only work for few ranks, ie 1d, 2d, and 3d. This
separates the lowering to normal conv and depthwise conv. The former one
is lowering to Linalg on tensors, and the latter one is leaving with the
same behavior, ie, lowering to Linalg on buffers.

Add support of TC conv ops to later GPU passes.

Depends on https://reviews.llvm.org/D96038