No deprecation warnings and no release asserts (#4965)
Fix release builds to build without asserts and turn off annoying
deprecation warnings.

The former is fallout from https://github.com/google/iree/pull/4836 and
the latter is something we've noticed being really annoying recently.

The setting of deprecation to no-error (explicitly inviting warning
spam) is something we seem to have inherited from Google's internal
build. My best guess is that people mostly only see these in the code
they're changing because everything is cached or something. If we want
to surface these warnings, we should set up clang-tidy. Spamming the
logs for anyone building anything in the project is not the way.