Tie operands and results when creating DispatchWorkgroupsOp (#5102)
For each resultant tensor, DispatchWorkgroupsOp now can indicate
to which operand tensor it ties to, in order to share the underlying
buffer storage. This can be used for subtensor_insert ops, where
we insert into the target tensor and also return the target
tensor as the result, which can happen as in-place update on
the buffer down the layers.