Adding a static library loader. (#5480)
This allows applications to inject a set of static libraries that can be resolved at runtime by name. This only includes the runtime portion: there's nothing in the compiler actually generating static libraries yet.

Future improvements to the driver API as part of #4298 will make it easier to inject these loaders.

Fixes #4679.