Fix pointer size detection (#5562)
This wasn't working on our internal 32-bit arm architecture for some
reason. Unlike many methods one can find on the internet, the one used
here doesn't require listing every possible architecture. It has the
disadvantage (like the one used previously) that it requires uintptr_t,
which is not technically required by the standard and it requires that
`uintptr_t` be the size of a void pointer, which is also not
technically required. Both these properties are checked at compile time
though, so if we ever have to deal with some weird architecture where
that isn't true, I guess we can find some other way.