Add barrier after duplicated dispatches for benchmarking (#5919)
Add a barrier after each clone. If the original dispatch has a small
problem size, simply duplicating without barrier will increasing the
number of subgroups and thus "help" filling the GPU. In the end we
will have an over optimistic result. Put in barriers avoids that,
but it assumes the input model is linear.

With this, the reported number reflects the kernel execution time
for MobileNetv2 with

```
iree-translate ... --iree-hal-benchmark-dispatch-repeat-count=32
iree-benchmark-module ... --batch_size=32
```

This is a step towards https://github.com/google/iree/issues/5248.
Next is working on duplicating the command buffer submission
as a whole.