Implement direct lowering of broadcasts from chlo.
* This is step 1 of the larger plan to do a one step HLO->IREE
conversion as part of the input pipeline. We are choosing to directly
legalize broadcasts at a higher level where we have more context to
resolve dynamic ambiguities.
* Introduced via a pass option that is used for testing the new behavior
and will be flipped as part of the overall switch, which is a few PRs
down the line.
* This uncovered several missing legalizations (now nicely flagged
because of enforcing a full conversion of the mhlo/chlo dialects) for
shift left/right.
* Also, broadcast_select semantics were changed recently and this only covers the
old forms (new forms are in the missing_legalizations.mlir) test and
will be implemented in a follow-up.
* Still iterating on some expanded lit tests but what is there is ready
for review.