Relicense IREE under the LLVM license (#6020)
To facilitate easy movement between IREE and upstream LLVM, we are
relicensing under the same license: the Apache 2.0 License with LLVM
Exceptions (https://llvm.org/LICENSE.txt, SPDX-License-Identifier:
Apache-2.0 WITH LLVM-exception).

In addition, we are moving to a Copyright header format that attributes
copyright to "The IREE Authors" and uses a separate AUTHORS file to
keep track of contributions. For this PR, I am just listing Google, but
I will send follow-ups to list Fraunhofer IML and Nod Labs (keeping it
separate so I can get those folks approval for the precise wording).

Note that the change to the header format is to recognize the valuable
contributions from our external collaborators and as always copyright is
retained by the author.

See the announcement at
https://groups.google.com/g/iree-discuss/c/NJohLQ5rOBg

The script I used to update (almost) all the headers is at
https://gist.github.com/GMNGeoffrey/8e85e1a088f59822ab61eff01f8d4ef6

The updates from that script are all in one commit. I recommend
reviewing the other commits separately from that one. You can also see
the raw diff from that change at
https://github.com/google/iree/commit/bdb5800ea51.diff

As part of this, I had to manually fiddle with ipynb files (actually
that was the most time-consuming aspect). I changed the format of the
license headers there to just be a block of markdown text rather than
something that was for some reason executable. Also had to manually give
the cells IDs, which hopefully will have worked.

To resolve the inevitable merge conflicts, I will likely drop the
changes from the script and rerun.

Tested:
```
$ grep -r --exclude-dir=third_party "Google LLC" . | grep -v website/site
./AUTHORS:Google LLC
```