Add scripts to filter and execute benchmarks on Android devices (#5999)
In IREE, we compile models/ops towards a specific CPU/GPU architecture.
When benchmarking those compiled models/ops, we need to excute them
on a concrete device. The device may host different CPU/GPU products.
So we need to filter the compiled benchmark suite according to the
specific device and only execute suitable ones on the device. This
helps to make benchmarking scalable: we don't need to know about the
specific phone in some centralized place.

This commit adds logic to probe the target Android device's CPU/GPU
characteristics using `adb` commands. Then it scans the benchmark
suite to find suitable ones and then push them to the device for
execution. The execution result is written as JSON files, with all
the details.

Right now this assumes the benchmark suite is of a particular structure:
`tf_models/<flavor>/<iree-driver>__<benchmark-flavor>__<target-arch>`.
This is similar to what is generated by the existing Python testing
utils. It is a reasonable assumption for now because these scripts
is just part of the whole benchmark pipeline; we cannot really define
and connect the final "exchange format" without having all the bits
ready. This can be easily adapted later for whatever final format we
settle down.