Create or update comments for benchmark results (#6145)
This is a follow up to https://github.com/google/iree/pull/6124,
where we categorized and sorted benchmark result tables.
But it's still noisy given we are posting many abbreviated
summary tables. This tries to reduce that by trying to find
the last previous comment from the bot's account and
update that.

Also, using issue API endpoints instead of pull request ones,
in the hope to avoid putting the bot as a "reviewer".