Add TranslationInfo Attribute. (#6175)
The decision of the lowering pipeline to use is currently returned as
a value to the caller. This isnt scalable since it would not handle
cases of multiple entry point functions as cleanly. Instead add
another attribute that the configuration selection should add on the
hal.executable.entry_point op to defines the pipeline to use for
lowering as well as the workload per workgroup value.

Some cleanups to use workloadPerWorkgroup instead of workgroupSize
which has overloaded meaning.
Also on the CPU side, only the root operation needs a lowering
configuration. Eventually that should be true for all backends since
the root op should drive the compilation.