Implementing support for device/executable multi-targeting. (#6456)
This begins the move to the HAL transformation pipeline using info
in the IR instead of in flags. More plumbing is required and all existing
flags still work (-iree-hal-target-backends, etc).

The major change here is that now `#hal.device.target` attributes are
attached to modules (or funcs) to denote which devices are being targeted.
Each of the target device attributes specifies the target backend and
backend-specific options for how to translate executables in the form
of `#hal.executable.target` attributes. The IR then has the required
information for runtime specialization and executable multi-targeting to
replace the global options used previously (when all this is done).

Future work will focus on moving the global flags for each backend into
the attributes, splitting the device backend from the executable backend,
and improving the code paths used for runtime selection.

Multiple simultaneous target devices are not supported but the
groundwork is going in. Much more work on environment specification will
be required and that is better done together with designing the flow-level
interactions.