Fix cuda hal for `-Wunused-but-set-parameter` (#6610)
This is a new check in clang that is going to be enabled globally
internally soon. It's not yet available in a released clang. If it ends
up being annoying, we could disable it for all of IREE internally. We
could also do that now.

Regardless, I think this fix is probably good, since we've identified
the issue. This is currently the only failure in IREE core.