Use IREE submodule dependencies in TF integrations build (#6817)
Instead of using all of TF's configuration of dependencies, we let
IREE's take precedence. This results in less surprising behavior when
editing IREE submodules and expecting a change to the integrations build
at the cost of building TF in a configuration more different from the is
one they use.

This didn't used to be possible because none of TF was configured with
`maybe`, but https://github.com/tensorflow/tensorflow/commit/7ed9c675
added similar functionality to `tf_http_archive`. This was more recently
blocked on TF's incompatibility with the upstream build files, but with
the removal of unnecessary references to
[`mlir:subpackages`](https://github.com/tensorflow/tensorflow/commit/417639263326)
and
[`mlir:friends`](https://github.com/tensorflow/tensorflow/commit/400d63c99ffc)
that issue should go away.