Remove the patterns that create dispatch regions without tile+fuse+distribute. (#7306)
With TiledOpInterface most of the operations to be executed on the
device can be tiled and distributed. So the pattern that created a
non-tiled and distributed dispatch region can be removed.
Simplify the code that decides which ops are root ops.

To make all tests pass, the last operation that needs to be tiled and
distributed is tensor.extract_slice. Add an external model to
implement the tile+distribute of this op to allow deleting the fall
back path.