[spirv] Revamp cooperative matrix support (#7332)
This pass splits the cooperative matrix support logic in the
SPIR-V backend to its own pass pipeline. This helps to isolate
the functionality and keep both worlds clean: we can have
proper dedicated handling of cooperative matrix and also avoid
lots of special checks in the normal pipeline.

The pass pipeline consists of the following new changes/passes:

* `NVIDIAConfig` is revamped to better find the supported
  cooperative matrix sizes.
* `SPIRVTileAndVectorizeToCooperativeOps`: tiles to subgroups
  and then performs vectorization to vector sizes matching
  by native cooperative matrix sizes. This pass basically takes
  its functionality from the existing `SPIRVTileAndDistribute`
  and `SPIRVVectorize`, with a curated list of patterns and
  configurations.
* `SPIRVVectorToCooperativeOps`: this takes the main patterns
  for the previous `SPIRVVectorToCooperativeMatrix` pass to
  lower vector transfer and contract ops, while removing all
  its cooperative applicability analysis. We additionally need
  to support splat cooperative matrix and elementwise ops there.

An example is checked in as tests to show that we can lower
matmul with consumer elementwise ops. The limitation is that
we are not using workgroup memory. That can be enabled later.

This commit copies various pieces from existing passes. Later
commits will clean up them from the main pipeline properly.