Remove flatbuffer from system library loader. (#7411)
This replaces them with a small suffix struct only used when additional
debug databases are required (like pdbs). The goal is to not have any
additional information required at runtime that is not queried from the
executable itself so the flatbuffer was unneeded, and now it's much
easier to get libraries out of vmfbs as the files inside are directly
usable by tools like executable_library_benchmark and objdump.