Enable benchmarks from TFLite flatbuffers on benchmark pipeline (#7494)
This replaces the three benchmarks derived from TFLite flatbuffers to
use the original flatbuffer source rather than a pre-compiled MLIR
file. These benchmarks are updated in place since the original source
is exactly the same. Using flatbuffers avoids a few issues:
1. The MLIR IR formats do not offer any backward compatibility
   guarantees. When MLIR syntax changes in a backward-incompatible way,
   the benchmark pipeline breaks.
2. Updates to the frontend compilation are not captured in real time.
   They are only incorporated as part of the updates above, which means
   that they are realized only when updating the MLIR files due to
   syntax changes.

This adds a dependency on the TFLIte frontend to benchmark generation.
In the automated pipeline, the additional Bazel compilation uses the
remote cache and should therefore usually be fast (as it shares the
cache with the normal CI run). In local generation, users will have to
either build `iree-import-tflite` from source or fetch it as part of a
release. Given that we would like to move all benchmarks to originate
from stable formats, I've made trying to run a TFLite benchmark without
`IREE_BUILD_TFLITE_COMPILER` set an error rather than a warning that
skips the benchmark. This explicit check just fails slightly earlier
and more clearly, as the `TARGET_FILE` lookup for `iree-import-tflite`
would fail subsequently.