Add a pattern to vectorize tensor.pad with conditional reads (#8352)
This commits adds a pattern to vectorize tensor.pad ops with
static result shapes. It generates `scf.if` ops to guard against
vector.transfer_read ops to make sure they are in bounds.

This is needed for generating better code after fusing pad
with its consumers. There are a few patterns to vectorize pad
upstream, but none of them generates the IR pattern needed for
mobile GPUs. Before we evolve upstream to the point of
supporting the case here, use this pattern in IREE.