Set configuration using untiled operations. (#8446)
With tile and distributed being moved out of Flow, the backends can now use the untiled operations to decide configuration. Make that change and move the TileAndDistributeToWorkgroups Pass to run after configuration specifications on all backends. This makes the `getUntiledShape` and `getUntiledResultShape` methods unnecessary and are removed. The `LoopTileAndDistributionInfo` is still needed since its load bearing for some `affine.min` canonicalizations on the SPIR-V side.
- On the CPU side the configuration setting made heavy use of the existence of tiled loops to decide configuration in an op agnostic way. Adapt the configuration selection to not rely on these tiled loops and use the untiled operations directly.
- One of the CUDA tests checking for illegal configuration seems to be off and is being looked at in #8456 . Disabling that test now, but will have to be fixed after the fact in that PR.

Co-authored-by: Lei Zhang <antiagainst@google.com>
Co-authored-by: Thomas Raoux <thomasraoux@google.com>