[spirv] Fuse tensor.pad ops into their consumer ops  (#8049)
It's typical for convolutions to require padding. Right now we handle
such padding with two separate steps, first performing `vkCmdFillBuffer`
to initialize the target buffer and then launch a specific kernel to
copy over the source content. These two steps also means two additional
pipeline barriers to sequence the execution.

This commit instead enables fusion of tensor.pad ops into the consumer
ops directly to avoid generating the buffer fill and copy. This is
expected to improve performance on mobile GPUs.

After fusing the tensor.pad op, we generate the kernel to contain both
a fast and a slow path. The fast path are for inner tiles that do not
need padding; the slow path is for boundary tiles that do need padding.
Both paths are vectorized properly.