Replace llvm_unreachable by assert(false) or pass failures (#8521)
All of the uses were in iree/compiler or other compiler directories, so
none of them was plausibly sufficiently performance-critical to justify
the risk of having the control flow potentially reach a
llvm_unreachable. Some of them were not invariants at all and are turned into pass failures / op errors.