Add a pass to bufferize early copy only dispatches. (#8529)
For cases where the dispatches are copy only , i.e. data is transfered
from one interface binding to another, bufferizing early results in a
linalg.generic (i.e. a copy) operation in the dispatch. The backends
can use this to generate code. This is the first step in dropping
TiledOpInterface for tensor.insert_slice and
tensor.extract_slice operations. This commit also adds patterns to
fold these operations with the flow.dispatch.tensor.load and
flow.dispatch.tensor.store operations. Eventually these patterns
will be moved to canonicalizations when the TiledOpInterface
implementation for tensor.insert_slice and tensor.extract_slice
are dropped.

Fixes #8509