Rearranging iree_notification_t code to make it easier to follow. (#8877)
The mix of atomic/futex and pthreads operations made it hard to know
when the pthread mutex should be locked and when atomics were ok. The
disabled synchronization path that no-op'ed also obscured the intended
behavior. Now we have the divergent code split up more cleanly so that
we can more easily inspect it for correctness.

iree_notification_await is left shared across the implementations to
preserve program behavior (that the condition is not checked with
a lock held) though we could probably tweak that if we wanted. Erring
on the side of keeping things the same for now.

Integrates the fix from #8876 for a hang/deadlock when using pthreads.