Pull the manylinux2014 docker image into IREE. (#8941)
This was being managed in a separate repo and used to have some exotic
things that needed to be done to make bazel work. Now, though, it is
formulamatic, so we just include it in our normal setup.

This also moves the TBB library install into the docker build (it was
being done at release time) in order to satisfy Tracy deps.