NFC refactor dispatch region formation to allow multiple results. (#8937)
Current dispatch region formation does not handle cases where the
fused operations themselves have uses outside of the ops pulled into
the same dispatch, i.e. it only handles cases where the results of the
dispatch are same as the results of the root op. Relax this to allow
for fusing with operations, where even the fused ops can have uses
outside of the dispatch. Enable this in a following commit.