Fix dana uploading failure due to duplicated numbers (#9061)
Right now we have two buildkite pipelines uploading benchmark
samples to Dana---one for Android targets and one for x86_64
CPUs. Each pipeline will try to create a new build for the
current landed commit. We need to allow overriding to make
sure both of them are accepted. (It should be the exact same
commit so no harm to override.)