Promoting the i64 VM extension to core. (#9138)
This unconditionally enables use of i64s in bytecode by moving all of
the ops into the core table. This has the benefit of reducing binary size
vs. when optional as the op table is much more dense.

This is a breaking change to bytecode and the major version is bumped.
Old vmfbs will fail to load and need to be rebuilt.