Only enable padding for linalg.matmul cases. (#9123)
There is an order for doing padding correctly:

1. Tile parallel loops
2. Pad operations in top-down order (i.e., topology order) for all parallel dims.
3. Tile reduction loops
4. Pad operations for all reduction dims. (In IREE's use case, there is only one operation that has reduction loops.)

Specifying the operation to pad is not easy today. We should have better
scheduling to do it. Either in transform dialect approach or creating a
new pass. This PR temporarily disables padding for other cases (e.g.,
matvec, batch_matmul, etc) because IREE puts more focus on matmul cases
for now.