Enabling split rodata blobs in VMFB archives. This changes the VMFB files to contain a length prefix so that the end of the module flatbuffer can be calculated and trailing rodata can be located. With this the module flatbuffer is still limited to 2GB but any amount of data may follow as all offsets are tracked as 64-bit values.
In order to make this work the polyglot zip emitter had to be reworked
as to not mess with the required rodata alignments. Since big file-like
blobs that would show in the zip are now stored outside the flatbuffer
there's also no need to do the crazy embedded zip headers and we can now
put a local file header in front of the module flatbuffer and show that
in the zip. A few bugs and compatibility issues were worked out and now
the vmfb files can be directly detected as zips by `file` and other tools
and everything I've tried (`unzip`, `7z`, Windows, etc) can open them.