Allow dispatch region formation to fuse producer and consumer when producer has multiple uses. (#8970)
Current heuristic to create dispatch regions has checks to fuse
producer with consumer only when the producer has a single use. This
is pessimizing. The only real requirement is that the consumer
dominates all the producers users. Use dominance info to relax this
requirement.

All the models we track currently though dont seem to benefit with
this change (they dont degrade either). The MobileBertInt8 model has
number of dispatches reduce from 1520 -> 1496.

Since the backends still have some issues to allow multi-result dispatches
to be compiled properly (thats WIP), this change is guarded by a flag
`--iree-flow-enable-multi-result-dispatches`.