Remove literal interpolation in Buildkite bootstrap pipeline (#9208)
Apparently when interpolated as the default, variables are typed as
strings, which is a bit confusing to me. We were getting
[422 errors](https://buildkite.com/iree/postsubmit/builds/412) in
the Postsubmit pipeline, which helpfully didn't include any error
message. I was able to get a more helpful message by trying to manually
update the steps in the UI.

Tested: Ran the API call to update the pipeline manually.