Demo adding a new pipeline (#9221)
Demo of how to add a new pipeline and a test that it works. Note that
the new test-pipeline configuration runs as part of the "unregistered"
pipeline, since it's not yet registered with Buildkite. After
submission, I'll need to mark the newly-created pipeline as public and
give access to run it on presubmit, as noted in
https://github.com/google/iree/tree/5d7dac9b3c/build_tools/buildkite#making-pipeline-public

Successful run: https://buildkite.com/iree/presubmit/builds/136