Use upstream vectorize transform op (#9412)
This is a functional change due to the differences in the implementation
of the upstream vectorize op and its local equivalent. Specifically, the
upstream op can be only targeted at an isolated-from-above payload IR
op, as opposed to the local equivalent that could be targeted at any
payload IR op or untargeted. Targeting an arbitrary op for vectorization
is problematic because vectorization is implemented as a set of
patterns, some of which perform vectorization and some of which are
enabling or cleaning; targeting an individual payload IR (implemented by
outlining the op) hinders the application of the latter class of
patterns. This is most visible in the double-tiling.mlir test, where the
local approach was not actually vectorizing `tensor.pad` operations
despite the corresponding option being set.