Adding --list_drivers, --list_devices, and --dump_devices flags. These allow for querying of registered drivers, all devices or devices for a particular driver, and detailed per-device information provided by the driver implementations. The intent is that drivers can dump relevant information like supported device features, limits, etc ala something like nvidia-smi or vulkaninfo.
CUDA and Vulkan drivers have gained the ability to parse some device
paths. CUDA can now either take the UUID/MIG or device ordinal and Vulkan
currently can only take a device ordinal. With the URI scheme this means
these all work to reference the same device:
```
--device=cuda://GPU-754d9ae2-8df5-f8e3-3502-182434a12876
--device=cuda://0
--device=cuda:0
```
(the UUID and ordinal are the same as printed by `nvidia-smi -L`)

Drivers can have arbitrarily complex paths and we can continue to add
support over time. The new listing commands act as the source of truth
for what devices are supported and how to reference them.

Progress on #9343.