Rename "translate" to "compile" across build systems. (#9442)
All uses of `iree-translate` were replaced with `iree-compile` and the `iree-translate` tool was removed. We still have some "translation passes" and "translation info" names in the compiler, but most language can be updated to "compile" and "compilation".

I'm separately planning on updating uses of `--iree-mlir-to-vm-bytecode-module` and related to flags to use the newer `--output-format` and `--compile-mode` flags (or just trusting the default values).