[LLVMCPU] Move GEMM based dispatches to use padding approach. (#9433)
For padding approach, the order is

1. Tile and fuse parallel dims
2. Pad operations for parallel dims in top-down order.
3. Tile reduction dims
4. Pad operations for reduction dims. In IREE, only root op could have reduction dims.

This PR adds more options to control the pipeline more general, and enable padding for all GEMM based dispatches, e.g., batch_matmul.

The current matching behavior is looking into op name. This blocks padding for `generic with reduction` + `generic` cases. We'll be able to enhance the matching when moving to transform dialect. In current approach, we use `setAnchorOpToRootOp` option to determine the target op when padding operands for reduction dims.