Fix float->[u]int conversion (#9602)
The current code is actually exercising undefined behavior. `lroundf`
can't be used for converting to uint32 because `long int` is only
guaranteed to be at least a 32-bit signed integer. The floats used in
the tests are also ones that can't be represented precisely in an f32.

Note that the f32->ui32 test was actually explicitly *testing*
overflow previously, as it was asserting that the rounded value ended
up being zero.