Removing iree/base/logging.h. (#9605)
Besides a few random spots it cropped up in the runtime/python bindings
the C++ logging code was only used in tests and tools where the
features provided for verbosity control aren't meaningful. Instead of
porting it all to C I've just gone and deleted it -
`fprintf(stderr)`/`std::cerr` is almost always what's wanted instead
and is much simpler to understand. As a library we try not to have any
log spew, tests can use gtest logging, and the tools are almost all
LLVM-based and use the logging functionality already available in
that world.

Fixes #2843.