Bump lint actions to use Ubuntu 20.04 and clang-format-12 (#9484)
A redo of #9473. I think it makes sense to run our lint actions at the
latest version. We only don't set it to latest due to
[Google security policy](https://opensource.google/documentation/reference/github/services#gha-bestpractices).

Tested: created commits in this PR with changes that should be caught
by each respective lint check and confirmed that the errors were as
expected.

Fixes https://github.com/google/iree/issues/9154