Make release run on self-hosted GitHub runners (#9432)
We currently have three n1-standard-64 runners for releases.

This dramatically reduces runtime for the linux releases, leaving the
macos releases as the bottleneck (still taking 3h for the compiler
package). We could consider separating these so mac releases are less
frequent. Here's a comparison of the runtimes (in minutes):

| Build                    | managed | self-hosted |
|--------------------------|---------|-------------|
| main-dist-linux          | 146.5   | 14.6        |
| py-compiler-pkg          | 177.1   | 25.0        |
| py-runtime-pkg           | 13.9    | 9.5         |
| py-tf-compiler-tools-pkg | 242.6   | 23.8        |

https://github.com/google/iree/runs/6905353191
vs
https://github.com/google/iree/runs/6810296844

Was previously #9430, but needed to be a branch on the main repository
to trigger the action on it for testing.