Fix compile errors in atomics_disabled (#9782)
iree_atomic_store_intptr(x, y) expanded to

(intptr_t) * (iree_atomic_int32_t*) (x) = (y)

The (intptr_t) cast on the left hand side had higher precedence than the
assignment operator so the entire expression

  (intptr_t) * (iree_atomic_int32_t*) (x)

was interpreted as a lvalue being assigned to, and that resulted in
compilation error. This showed up only on riscv CI on #9777, presumably
because it's the only place where we use atomics_disabled, and on that
PR, presumably because it introduced a use of iree_atomic_store_intptr.