Add a workflow for building and testing the runtime post-submit (#9774)
This uses GitHub's hosted runners, since they are sufficient when only
building the runtime. It mostly a step towards migrating our initial
Buildkite workflows and finding parity with those. Note that this is
not doing anything to make the entire workflow idempotent. We will
likely need to move to our own artifact storage scheme (backed by GCS)
to make that work. We definitely incur a cost from having to upload and
download the build directory rather than building and testing in the
same directory (1.5 minutes to upload. download is negligible), but I
think the modularity is worth it and we can find a way to speed it up.